backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0
geoalchemy2>=0.14
psycopg2-binary>=2.9
pandas>=2.1
openpyxl>=3.1
numpy>=1.26
pulp>=2.8
highspy>=1.7
requests>=2.31
pydantic>=2.5
python-dotenv>=1.0
